move to ragas_examples
