+++ examples/ragas_examples/__init__.py
@@ -1,0 +1,16 @@
+"""
+Ragas Examples Package
+
+This package contains official examples demonstrating how to use Ragas for evaluating 
+different types of AI applications including RAG systems, agents, prompts, workflows, 
+and LLM benchmarking.
+
+Available example modules:
+- agent_evals: Agent evaluation examples
+- benchmark_llm: LLM benchmarking and comparison examples  
+- prompt_evals: Prompt evaluation examples
+- rag_eval: RAG system evaluation examples
+- workflow_eval: Workflow evaluation examples
+"""
+
+__version__ = "0.1.0"
