+++ examples/ragas_examples/prompt_evals/evals.py
@@ -1,4 +1,4 @@
-from prompt import run_prompt
+from .prompt import run_prompt
 
 from ragas import experiment
 from ragas.experimental import Dataset
